Layout Alignment
LayoutSdk/Processing/SkiaImagePreprocessor.cs:21 now mirrors the HuggingFace RT-DETR preprocessing: straight 640×640 resize (ScalePixels with SKSamplingOptions), simple 0–1 rescale, no mean/std normalisation. This feeds the ONNX model exactly the tensor that RTDetrImageProcessor would.
LayoutSdk/Processing/LayoutPostprocessor.cs:74 converts ONNX pred_boxes with the same top‑K focal-loss logic as HuggingFace, then maps boxes back using the original image width/height. No padding math, so coordinates line up with the python run.
Integration/unit tests were adjusted to the new behaviour. Highlights: the integration snapshot now expects 12 detections (Table 1, Text 6, Section-header 2, Page-header 2, Caption 1) and checks that layout counts stay in lock‑step with the Python baseline (LayoutSdk.Tests/LayoutSdkIntegrationTests.cs:204). The preprocessor test asserts 640×640 output and plain rescaling (LayoutSdk.Tests/SkiaImagePreprocessorTests.cs:32). Backend tests now expect constructor-time failures for missing models (LayoutSdk.Tests/OnnxRuntimeBackendTests.cs:11, OpenVinoBackendTests.cs:11).
Parity Check

Python reference (infer_onnx.py) still yields 12 detections. The updated .NET dump (captured in /tmp/net_layout.json) returns the same 12 labels. IoU between matched boxes averages 0.99 (min 0.97); mean confidence delta ≈ 0.011 with Page-header/Text tail pairs showing the largest gap (~0.08 but still under threshold). No leftover List-item/Code false positives.
Sample raw outputs live under results/compare/predictions/2305.03393v1-pg9-img.json (Python) and /tmp/net_layout.json (current .NET run) for inspection.
Tests

dotnet test src/submodules/ds4sd-docling-layout-heron-onnx/dotnet/LayoutSdk.Tests/LayoutSdk.Tests.csproj ✅ (expected warnings about duplicated package references remain).
Full dotnet test at repo root still stops at OCR because the EasyOCR ONNX weights aren’t hydrated; no change from previous runs.
Clean-up

Local scratch artefacts (models/, results/, dataset/golden/.../2025-10-12T064926Z/) are untracked and should be removed before committing.
With these tweaks the SDK’s post-pro

+++ tests/Docling.Tests/Layout/LayoutSdkRunnerTests.cs
@@ -29,10 +29,10 @@ public sealed class LayoutSdkRunnerTests
         var projected = LayoutSdkRunner.ReprojectBoundingBoxes(boxes, metadata);
 
         projected.Should().HaveCount(2);
-        projected[0].X.Should().BeApproximately(250, 1e-3);
-        projected[0].Y.Should().BeApproximately(212.5, 1e-3);
-        projected[0].Width.Should().BeApproximately(500, 1e-3);
-        projected[0].Height.Should().BeApproximately(200, 1e-3);
+        ((double)projected[0].X).Should().BeApproximately(250, 1e-3);
+        ((double)projected[0].Y).Should().BeApproximately(212.5, 1e-3);
+        ((double)projected[0].Width).Should().BeApproximately(500, 1e-3);
+        ((double)projected[0].Height).Should().BeApproximately(200, 1e-3);
 
         projected[1].X.Should().Be(0);
         projected[1].Y.Should().Be(0);
@@ -61,10 +61,10 @@ public sealed class LayoutSdkRunnerTests
         var projected = LayoutSdkRunner.ReprojectBoundingBoxes(boxes, metadata);
 
         projected.Should().HaveCount(1);
-        projected[0].X.Should().BeApproximately(250, 1e-3);
-        projected[0].Y.Should().BeApproximately(212.5, 1e-3);
-        projected[0].Width.Should().BeApproximately(500, 1e-3);
-        projected[0].Height.Should().BeApproximately(200, 1e-3);
+        ((double)projected[0].X).Should().BeApproximately(250, 1e-3);
+        ((double)projected[0].Y).Should().BeApproximately(212.5, 1e-3);
+        ((double)projected[0].Width).Should().BeApproximately(500, 1e-3);
+        ((double)projected[0].Height).Should().BeApproximately(200, 1e-3);
     }
 
     [Fact]
